Pull `transformQueryParams` out into a utility function
